Working version using standard atan2 function

+++ Segment.cpp
@@ -37,16 +37,16 @@ float Segment::getLength(){
 }
 
 float Segment::angleBetween(const Segment &other) const {
-    float x1=B.getX()-A.getX();
+    short x1=B.getX()-A.getX();
     //float x2=other.B.getX()-A.getX();
-    float x2=other.B.getX()-other.A.getX();
-    float y1=B.getY()-A.getY();
+    short x2=other.B.getX()-other.A.getX();
+    short y1=B.getY()-A.getY();
 
     //float y2=other.B.getY()-A.getY();
-    float y2=other.B.getY()-other.A.getY();
+    short y2=other.B.getY()-other.A.getY();
 
-    float dot = x1*x2 + y1*y2;//      # dot product between [x1, y1] and [x2, y2]
-    float det = x1*y2 - y1*x2;//      # determinant
+    int dot = x1*x2 + y1*y2;//      # dot product between [x1, y1] and [x2, y2]
+    int det = x1*y2 - y1*x2;//      # determinant
     float alfa = atan2(det, dot);
     return alfa;
